pyproject.toml
[tool.poetry]
name = "gpu-speech-metrics"
version = "0.1.0"
description = ""
authors = ["Koen Oostermeijer <koen@oostermeijer.net>"]
readme = "README.md"
packages = [{include = "gpu_speech_metrics"}]

[tool.poetry.dependencies]
python = "^3.12"
datasets = "^3.6.0"
numpy = "<=2.0" # Needed for dataloading
torch = "^2.7.1"
torchaudio = "^2.7.1"
torchvision = "^0.22.1"
transformers = "^4.52.4"
pkginfo = "^1.12.1.2"
onnx2torch = "^1.5.15"
scipy-stubs = "^1.15.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.5.1"
pesq = "^0.0.4"
torch-pesq = "^0.1.2"
torchtyping ="^0.1.5"
pystoi = "^0.4.1"
torchmetrics = "^1.7.3"
librosa = "^0.11.0"
pre-commit = "^4.2.0"
ruff = "^0.12.3"

[tool.ruff]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "__pypackages__",
    ".pytest_cache",
    ".vscode"
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "B"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
